Move ValidateCustomLess to SettingsServiceProvider.  Use existing convention for validator.
